refactor and formatting
